Convert all pages to php
Copied the website folder to websiteCopy and made edits in there in case
I messed everything up.

Changed all html pages to php so we can work with php on all pages and
made a template for productpages in php.

Added counter to count how many items are in your cart based on cookies.

Edited to config.php to be able to create database connection object.

Fixed problem with drop down menus not working on product pages.

Changed "Sign In" on the Register pop up to "Register" since all it does
after clicking that button is redirect you to the index page.

Made it so you can add stuff to cart from the thumbnail page.

Made it so the pages retrieve all the informatino from the inventory
table. EX: The thumbnails, name, price, etc. Also does this depending on
if it is for male or female so that the female shirts do not show up on
the male shirt products list. This makes it so all we need to do is get
an image, create thumbnail of image using:

http://makethumbnails.com/#dropzone

Put thumbnail in thumbnails image folder. create product page from
template. Then add the quantity instock, price, gender(male, female),
product_id (needs to be the next number after the most recent product_id
or it will start to get confusing), pictureLink (ex.
images/thumbnails/white-shirt_tn.jpg), link to product(ex.
www.techsavvyclothing.com/distwebfall2016/website/white-shirt.html),
productName(ex. M/ Basic White Tee), code (which will be the same as the
products .php file name without the .php extension: ex. white-shirt). It
will then automatically populate the product lists.

TO DO STILL:

Need to figure out how to limit the number of items displayed on the
"View Selections" page.

Need to make the button on the products page to add to cart.

Need to mess with the css to get it to look like it used to or close to
it.

Need to figure out why sign in comes up when the signout button is
pressed.  And add something that clears the cart when you sign out.

Need to figure out how to start getting into check out.

Need to figure out the search function or if we just want to get rid of
it honestly.

I know I probably have forgotten some things so we can talk about next
time we meet.